requirements.txt
asgiref==3.2.10
cffi==1.15.1
cryptography==39.0.1
Django==3.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
